Add files via upload
The previous script has been completely revamped and refactored to include a default file that can be accessed by the main WebRequest2.1.py. This upgrade includes error handling, a fallback function in case PowerShell fails during the web request process, and plenty of room for further improvements. 

The cong.ini file has been designated as the [Default] WebRequest.py. By changing the output file path, you can now generate a .HTML file of the requested URL. 

In the future, updates will include file name generation and editing in the configuration file, instead of having to manually add the name of the output file at the end of your output_directory.